+++ src/sites/link/ohleech.com.js
@@ -1,9 +1,0 @@
-_.register({
-  rule: {
-    host: /^(www\.)?ohleech\.com$/,
-    path: /^\/dl\/$/,
-  },
-  async ready () {
-    $.window.startdl();
-  },
-});