+++ src/sites/link/sylnk.net.js
@@ -5,8 +5,6 @@ _.register({
     {
       host: [
         /^(www\.)?sylnk\.net$/,
-        /^dlneko\.(com|net|org)$/,
-        /^rumahsimpel\.com$/,
       ],
       query: /link=([^&]+)/,
     },
@@ -16,46 +14,17 @@ _.register({
       query: /v=([^&]+)/,
     },
     {
-      host: /^(www\.)?safelinkair\.com$/,
-      path: /^\/code$/,
-      query: /(?:\?|&)link=([a-zA-Z0-9/=]+)(?:$|&)/,
-    },
-    {
       host: [
-        /^link\.filmku\.net$/,
-        /^www\.healthygress24\.ga$/,
-        /^kombatch\.amankan\.link$/,
-      ],
-      path: /^\/p\/(go|healty-lie)\.html$/,
-      query: /^\?url=([a-zA-Z0-9/=]+)$/,
-    },
-    {
-      host: [
-        /^(gadget|auto|sports)14\.pw$/,
+        /^sports14\.pw$/,
         /^motosport\.pw$/,
-        /^nar-04\.tk$/,
         /^lindung\.in$/,
         /^motonews\.club$/,
         /^ww[23]\.picnictrans\.com$/,
-        /^gadget13\.com$/,
         /^azhie\.net$/,
         /^ww2\.awsubs\.co$/,
-        /^autorp\.us$/,
         /^plantaheim\.web\.id$/,
       ],
       query: /^\?d=([a-zA-Z0-9/=]+)$/,
-    },
-    {
-      host: /^www\.anisubsia\.tk$/,
-      path: /^\/p\/link\.html$/,
-      query: /^\?url=([a-zA-Z0-9/=]+)$/,
-    },
-    {
-      host: [
-        /^www\.insurance1\.tech$/,
-        /^www\.freeanimeonline\.xyz$/,
-      ],
-      query: /^\?site=([a-zA-Z0-9/=]+)/,
     },
     {
       host: /^i\.gtaind\.com$/,
@@ -72,12 +41,7 @@ _.register({
     },
     {
       host: /^sehatlega\.com$/,
-      query: /^\?lanjut=([a-zA-Z0-9/=]+)$/,
-    },
-    {
-      host: /^shorten\.id$/,
-      // note the trailing `=`
-      query: /^\?url=([a-zA-Z0-9/=]+)=$/,
+      query: /^\?r=([a-zA-Z0-9/=]+)$/,
     },
     {
       host: /^www\.compartiendofull\.net$/,
@@ -109,22 +73,13 @@ _.register({
         // safelinkreviewer.com
         // safelinkreview.co
         /^safelink(s?review(er)?)\.com?$/,
-        /^(getcomics|miuitutorial)\.gq$/,
-        /^awsubs\.cf$/,
-        /^awsubsco\.ga$/,
       ],
       query: /id=([\w\\]+=*)/,
     },
     {
       host: [
-        /^(www\.)?dlneko\.com$/,
-        /^(satuasia|tawaku)\.com$/,
-        /^ww3\.manteb\.in$/,
-        /^link\.filmku\.net$/,
-        /^www\.muucih\.com$/,
         /^(naisho|filmku|henpoi)\.lompat\.in$/,
         /^edogawa\.lon\.pw$/,
-        /^telolet\.in$/,
       ],
       query: /go=([\w\\]+=*)/,
     },
@@ -157,21 +112,29 @@ _.register({
 });
 
 _.register({
+  rule: {
+    host: /^(gameinfo|apasih)\.pw$/,
+    query: /^\?id=([a-zA-Z0-9/=]+)$/,
+  },
+  async ready () {
+    const f = $('form');
+    f.submit();
+  },
+});
+
+_.register({
   rule: [
     {
       host: [
-        /^(designinghomey|ani-share|sinopsisfilmku|autolinkach)\.com$/,
+        /^(ani-share|autolinkach)\.com$/,
         /^motonews\.club$/,
-        /^(autofans|landscapenature|apasih|gameinfo)\.pw$/,
-        /^(sidespace|erogedownload)\.net$/,
-        /^otoviral\.racing$/,
+        /^(autofans|landscapenature)\.pw$/,
         /^www\.lifesurance\.info$/,
       ],
       query: /get=([^&]+)/,
     },
     {
-      host: /^sipkur\.us$/,
-      path: /\.html$/,
+      host: /^(gameinfo|apasih)\.pw$/,
     },
   ],
   async ready (m) {
@@ -208,16 +171,6 @@ _.register({
     const a = $('div.kiri > center > a');
     await $.openLink(a.href);
   },
-});
-
-_.register({
-  rule: {
-    host: /^aw-games\.net$/,
-  },
-  async ready () {
-    const a = $('.iklan a');
-    await $.openLink(a.href);
-  }
 });
 
 _.register({